Build plugins in `pull_request` event instead of `pull_request_target`
This significantly complicates the deployment process, but is necessary
because workflows running in `pull_request_target` have write access to
the main repository, as well as access to secrets. A bad actor could
very easily simply print these to STDOUT or pipe them over to a remote
server without us noticing.

This is also a problem with the current CI workflow, but partially
mitigated through the need of approval before the plugin is even run.
However, one could very easily sneak this past the review process by
having this malicious code hidden in a dependency and/or NPM install
hook for example.

Due to the fact that the `pull_request` event doesn't have access to
secrets, we need a second workflow that picks up the result from the
`build_plugins` workflow and pushes it to the testing store.

Co-authored-by: TrainDoctor <traindoctor@protonmail.com>
Co-authored-by: AAGaming <aa@mail.catvibers.me>