Update ControllerTools to v1.2.1
Update ControllerTools to v1.2.1 and change the submodule to point to
the new repository. The alphamercury repository is no longer maintained.
You can verify this by going to the old url [1] and verifying that it
redirects to the new url [2].

This new version pulls in the latest dependencies and makes the plugin
use the global DFL to address an error [3] in the UI after the Steam Deck
client update.

[1] https://github.com/alphamercury/ControllerTools
[2] https://github.com/jfernandez/ControllerTools
[3] jfernandez/ControllerTools#43

Signed-off-by: Jose Fernandez <josef@netflix.com>