Merge PR's and minor updates
This mergest PR's from Beebles, GloriousEggroll and Ugzuzg.

Beebles: changed scripts to work on nixos
GloriousEggroll: Added changes to make Umu fixes work correctly.
Ugzuzg: Changed sorting to be case insensitive on the game grid screen.

Fix for Garden Story where ${} was in the name the $ needed escaping.

Added args that needs to be passed to games that was over looked in the initial builds (very few if any games might be affected by this).